1.0.1 — fix voice capture stopping early on pauses
Continuous dictation: keep the audio engine running across utterance
finalization and pauses; rotate recognition segments and accumulate text
so recording ends only on user stop. Preserve the last in-progress words on
stop/save, resume across call/Siri interruptions, guard against a wedged
recognizer, and fix a CGFloat/cos ambiguity that broke the build on newer Xcode.

+++ Sources/UI/ConstellationView.swift
@@ -33,6 +33,6 @@ struct ConstellationView: View {
     private func pos(_ i: Int, c: CGPoint, r: CGFloat) -> CGPoint {
         let n = max(top.count - 1, 1)
         let a = (Double(i - 1) / Double(n)) * 2 * .pi - .pi/2
-        return CGPoint(x: c.x + r * cos(a), y: c.y + r * sin(a))
+        return CGPoint(x: c.x + r * CGFloat(cos(a)), y: c.y + r * CGFloat(sin(a)))
     }
 }
